requirements.txt
# Base requirements
Django==1.10
django-dotenv==1.4.1

pandas==0.18.1
mysqlclient==1.3.7
requests==2.10.0

# Test requirements
coverage==4.2
factory-boy==2.7.0
flake8==3.0.3
pip-conflict-checker==0.3
